backend/requirements.txt
Django==4.2.27
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==11.3.0
gunicorn==23.0.0
